+++ src/monst.c
@@ -1977,6 +1977,7 @@ NEARDATA struct permonst mons_init[] = {
           | M1_CARNIVORE,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 20, CLR_CYAN),
+    /* Red dragons deal more damage with their breath */
     MON("red dragon", S_DRAGON, LVL(15, 9, -1, 20, -4), (G_GENO | 1),
         A(ATTK(AT_BREA, AD_FIRE, 6, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
           ATTK(AT_CLAW, AD_PHYS, 1, 4), ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK,
@@ -2014,16 +2015,19 @@ NEARDATA struct permonst mons_init[] = {
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 20, CLR_BRIGHT_MAGENTA),
     /* disintegration breath is actually all or nothing, not 1d255 */
+    /* black dragons butt with their horns. */
     MON("black dragon", S_DRAGON, LVL(15, 9, -1, 20, -6), (G_GENO | 1),
         A(ATTK(AT_BREA, AD_DISN, 1, 255), ATTK(AT_BITE, AD_PHYS, 3, 8),
-          ATTK(AT_CLAW, AD_PHYS, 1, 4), ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK,
+          ATTK(AT_BUTT, AD_PHYS, 1, 4), ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK,
           NO_ATTK),
         SIZ(WT_DRAGON, 1500, MS_ROAR, MZ_GIGANTIC), MR_DISINT, MR_DISINT,
         M1_FLY | M1_THICK_HIDE | M1_NOHANDS | M1_SEE_INVIS | M1_OVIPAROUS
             | M1_CARNIVORE,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 20, CLR_BLACK),
-    MON("blue dragon", S_DRAGON, LVL(15, 9, -1, 20, -7), (G_GENO | 1),
+    /* blue dragons butt and are faster than other types of dragon and butt with
+       their horn. */
+    MON("blue dragon", S_DRAGON, LVL(15, 12, -1, 20, -7), (G_GENO | 1),
         A(ATTK(AT_BREA, AD_ELEC, 4, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
           ATTK(AT_BUTT, AD_PHYS, 2, 4), NO_ATTK, NO_ATTK,
           NO_ATTK),
@@ -2032,6 +2036,7 @@ NEARDATA struct permonst mons_init[] = {
             | M1_CARNIVORE,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 20, CLR_BLUE),
+    /* green dragon breath leaves a poison trail, and they stalk their opponents. */
     MON("green dragon", S_DRAGON, LVL(15, 9, -1, 20, 6), (G_GENO | 1),
         A(ATTK(AT_BREA, AD_DRST, 4, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
           ATTK(AT_CLAW, AD_PHYS, 1, 4), ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK,
@@ -2039,7 +2044,7 @@ NEARDATA struct permonst mons_init[] = {
         SIZ(WT_DRAGON, 1500, MS_ROAR, MZ_GIGANTIC), MR_POISON, MR_POISON,
         M1_FLY | M1_THICK_HIDE | M1_NOHANDS | M1_SEE_INVIS | M1_OVIPAROUS
             | M1_CARNIVORE | M1_POIS,
-        M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
+        M2_HOSTILE | M2_STRONG | M2_STALK | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 20, CLR_GREEN),
     MON("gold dragon", S_DRAGON, LVL(15, 9, -1, 20, 4), (G_GENO | 1),
         A(ATTK(AT_BREA, AD_FIRE, 6, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
@@ -2061,7 +2066,7 @@ NEARDATA struct permonst mons_init[] = {
         0, 20, CLR_YELLOW),
     /* If you see an elder dragon, run. These monsters were added to increase
        challenge in the late game, and all of them are extremely nasty. */
-    MON("ooze dragon", S_DRAGON, LVL(25, 9, -10, 20, -7), (G_GENO | 1),
+    MON("ooze dragon", S_DRAGON, LVL(25, 9, -10, 20, -7), (G_GENO | G_PLANES | 1),
         A(ATTK(AT_BREA, AD_ACID, 4, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
           ATTK(AT_ENGL, AD_ACID, 3, 6), ATTK(AT_CLAW, AD_PHYS, 1, 4),
           ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK),
@@ -2070,7 +2075,7 @@ NEARDATA struct permonst mons_init[] = {
                   M1_AMORPHOUS | M1_CARNIVORE | M1_ACID,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 31, CLR_BRIGHT_GREEN),
-    MON("razor dragon", S_DRAGON, LVL(25, 18, -10, 20, 7), (G_GENO | 1),
+    MON("razor dragon", S_DRAGON, LVL(25, 18, -10, 20, 7), (G_GENO | G_PLANES | 1),
         A(ATTK(AT_BITE, AD_PHYS, 4, 6), ATTK(AT_CLAW, AD_PHYS, 10, 4),
         ATTK(AT_CLAW, AD_PHYS, 10, 4), NO_ATTK, NO_ATTK, NO_ATTK),
         SIZ(WT_DRAGON, 1500, MS_ROAR, MZ_GIGANTIC), MR_COLD, 0,
@@ -2078,7 +2083,7 @@ NEARDATA struct permonst mons_init[] = {
             | M1_METALLIVORE,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 29, CLR_CYAN),
-    MON("filth dragon", S_DRAGON, LVL(25, 9, -10, 20, -7), (G_GENO | 1),
+    MON("filth dragon", S_DRAGON, LVL(25, 9, -10, 20, -7), (G_GENO | G_PLANES | 1),
         A(ATTK(AT_BREA, AD_DRST, 4, 6), ATTK(AT_BITE, AD_PHYS, 3, 8),
           ATTK(AT_CLAW, AD_DISE, 1, 4), ATTK(AT_CLAW, AD_DISE, 1, 4),
           NO_ATTK, NO_ATTK),
@@ -2087,7 +2092,7 @@ NEARDATA struct permonst mons_init[] = {
                   M1_CARNIVORE | M1_POIS | M1_REGEN,
         M2_HOSTILE | M2_STRONG | M2_NASTY | M2_GREEDY | M2_JEWELS | M2_MAGIC,
         0, 31, CLR_BROWN),
-    MON("hex dragon", S_DRAGON, LVL(25, 9, -10, 80, -6), (G_GENO | 1),
+    MON("hex dragon", S_DRAGON, LVL(25, 9, -10, 80, -6), (G_GENO | G_PLANES | 1),
         A(ATTK(AT_MAGC, AD_SPEL, 4, 6), ATTK(AT_BITE, AD_CURS, 3, 8),
           ATTK(AT_CLAW, AD_PHYS, 1, 4), ATTK(AT_CLAW, AD_PHYS, 1, 4), NO_ATTK,
           NO_ATTK),